small treasury refactor, rename to transparency
