+++ apps/cave/components/Transparency/VestedTokensDialogs/VestedTokenDialog.tsx
@@ -3,7 +3,7 @@ import { Token } from '@concave/core'
 import { InfoIcon } from '@concave/icons'
 import { Button, Card, Flex, Spinner, Text, Tooltip } from '@concave/ui'
 import { ToggleButton } from 'components/ToggleButton'
-import { VestedTokenButtonProps } from 'components/Treasury/TreasuryRedeemCard'
+import { VestedTokenButtonProps } from 'components/Transparency/TreasuryRedeemCard'
 import { BigNumber } from 'ethers'
 import { formatEther, parseEther } from 'ethers/lib/utils'
 import { useEffect, useState } from 'react'
